prep-scripts/requirements.txt
smbus2
RPi.GPIO
unicornhat
